+++ client/src/pages/Login/Login.module.css
@@ -1,95 +1,94 @@
 .container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: #efefef;
-    gap: 20px;
-    width: 100%;
-    max-width: 500px;
-    margin: 20px auto;
-    padding: 30px 40px;
-    box-shadow: rgba(0, 0, 0, 0.16) 0px 10px 36px 0px,
-      rgba(0, 0, 0, 0.06) 0px 0px 0px 1px;
-  }
-  .logo{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-  }
-  .CompanyName{
-    text-align: center;
-    font-size: 36px;
-    font-weight: 600;
-    letter-spacing: 2px;
-  }
-  .title{
-    font-size: 1.6rem;
-  }
-  .form {
-    display: flex;
-    flex-direction: column;
-    gap: 30px;
-    width: 100%;
-  }
-  .group {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-  }
-  .group label {
-    font-size: 1.6rem;
-  }
-  .group input {
-    font-size: 1.4rem;
-    font-weight: 400;
-    width: 100%;
-    max-width: var(--form-max-width);
-    height: 45px;
-    padding: 5px 10px;
-    outline: none;
-    border: 1px solid var(--input-border);
-    border-radius: 6px;
-  }
-  .group button {
-    background-color: #138D75;
-    color: white;
-    width: 100%;
-    height: 50px;
-    border: none;
-    outline: none;
-    padding: 12px;
-    border-radius: 4px;
-    user-select: none;
-    box-shadow: rgba(0, 0, 0, 0.2) 0px 1px 3px 1px;
-    transition: all 0.15s ease-in-out 0s;
-    font-weight: 500;
-    line-height: 20px;
-    font-size: 22px;
-    letter-spacing: 2px;
-    justify-content: center;
-  }
-  .group button:hover{
-      opacity: 0.7;
-      border-radius: 4px;
-  }
-  .redirect{
-      align-self: flex-start;
-      display: flex;
-      flex-direction: row;
-      width: 100%;
-      gap: 20px;
-      font-size: 1.2rem;
-      padding-bottom: 20px;
-      border-bottom: 1px solid black;
-      font-size: 14px;
-      
-  }
-  .redirect a {
-      color: #138D75;
-      text-decoration: none;
-  }
-  .license{
-      display: flex;
-      text-align: center;
-      font-size: 14px;
-  }
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-color: #efefef;
+  gap: 20px;
+  width: 100%;
+  max-width: 500px;
+  margin: 20px auto;
+  padding: 30px 40px;
+  box-shadow: rgba(0, 0, 0, 0.16) 0px 10px 36px 0px,
+    rgba(0, 0, 0, 0.06) 0px 0px 0px 1px;
+}
+.logo {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+.CompanyName {
+  text-align: center;
+  font-size: 36px;
+  font-weight: 600;
+  letter-spacing: 2px;
+}
+.title {
+  font-size: 1.6rem;
+}
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+  width: 100%;
+}
+.group {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+.group label {
+  font-size: 1.6rem;
+}
+.group input {
+  font-size: 1.4rem;
+  font-weight: 400;
+  width: 100%;
+  max-width: var(--form-max-width);
+  height: 45px;
+  padding: 5px 10px;
+  outline: none;
+  border: 1px solid var(--input-border);
+  border-radius: 6px;
+}
+.group button {
+  background-color: #138d75;
+  color: white;
+  width: 100%;
+  height: 50px;
+  border: none;
+  outline: none;
+  padding: 12px;
+  border-radius: 4px;
+  user-select: none;
+  box-shadow: rgba(0, 0, 0, 0.2) 0px 1px 3px 1px;
+  transition: all 0.15s ease-in-out 0s;
+  font-weight: 500;
+  line-height: 20px;
+  font-size: 22px;
+  letter-spacing: 2px;
+  justify-content: center;
+}
+.group button:hover {
+  opacity: 0.7;
+  border-radius: 4px;
+}
+.redirect {
+  align-self: flex-start;
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  gap: 20px;
+  font-size: 1.2rem;
+  padding-bottom: 20px;
+  border-bottom: 1px solid black;
+  font-size: 14px;
+}
+.redirect a {
+  color: #138d75;
+  text-decoration: none;
+}
+.license {
+  display: flex;
+  text-align: center;
+  font-size: 14px;
+}
