+++ client/src/pages/index.js
@@ -1,1 +1,3 @@
 export { default as Login } from './Login';
+export { default as Register } from './Register/Register';
+
